initial manual command
