+++ main.go
@@ -3,11 +3,11 @@ package main
 import (
 	"os"
 
-	"github.com/AppleGamer22/cocainate/cmd"
+	"github.com/AppleGamer22/cocainate/commands"
 )
 
 func main() {
-	if err := cmd.RootCommand.Execute(); err != nil {
+	if err := commands.RootCommand.Execute(); err != nil {
 		// _, _ = fmt.Fprintf(os.Stderr, "%v\n", err)
 		os.Exit(1)
 	}
